requirements.txt
a2a-sdk[http-server]>=0.3.0
h2ogpte==1.6.55
python-dotenv==1.2.1
uvicorn>=0.30.0